textures and math
sickness got me for a couple days but i'm back

okay i finished the texture stuff but did NOT add the second texture in yet. I gave it an attempt but i just lost myself halfway through, so i'll do that again later down the line

did a bunch of the maths stuff. i'm using the github with the unit labs on it as a guide on what to complete for this assignment, but when i get to the doc to fill things out if it turns out to be useless i'll just remove it